+++ demos/chinook/src/main/java/is/codion/demos/chinook/domain/ChinookImpl.java
@@ -119,7 +119,7 @@ public final class ChinookImpl extends DomainModel implements Chinook {
 																		WHERE track.album_id = album.id"""),
 										Album.TAGS.define()
 														.column()
-														.columnClass(Array.class, new TagsConverter(), ResultSet::getArray),
+														.converter(Array.class, new TagsConverter(), ResultSet::getArray),
 										Album.RATING.define()
 														.subquery("""
 																		SELECT AVG(rating) FROM chinook.track